feat: implement task details modal with support for assignees, checklists, comments, and labels
