+++ src/components/project-tabs/TasksTab.tsx
@@ -734,6 +734,7 @@ export function TasksTab({ projectId }: { projectId: number }) {
         comments={taskComments}
         commentsLoading={commentsLoading}
         currentUserId={user?.id || ""}
+        boardMembers={boardMembers}
         onAddComment={handleAddComment}
         onDeleteComment={handleDeleteComment}
         onUpdateTask={handleUpdateTaskField}
